Refactor to export flags into a seperate file; renaming crdmanagement/crds.go to manager.go
